move avro parsing into a separate module
