+++ backend/src/lib/avro/mod.rs
@@ -1,3 +1,8 @@
 mod avro_parser;
 mod avro_to_json;
+mod helpers;
+mod json_to_avro;
 mod schema_provider;
+
+pub use avro_parser::AvroParser;
+pub use schema_provider::SchemaProvider;
