+++ backend/src/lib/parser/mod.rs
@@ -1,6 +1,4 @@
-mod avro_parser;
 mod record_parser;
-mod schema_provider;
 mod string_parser;
 
 pub use record_parser::{Parser, ParserMode, RecordParser};
